+++ src/api/groups/members/members.rs
@@ -25,6 +25,9 @@ pub struct GroupMembers<'a> {
     /// A search string to filter members by.
     #[builder(setter(name = "_user_ids"), default, private)]
     user_ids: BTreeSet<u64>,
+    /// Skip certain user results.
+    #[builder(setter(name = "_skip_users"), default, private)]
+    skip_users: BTreeSet<u64>,
 }
 
 impl<'a> GroupMembers<'a> {
@@ -51,6 +54,25 @@ impl<'a> GroupMembersBuilder<'a> {
         self.user_ids.get_or_insert_with(BTreeSet::new).extend(iter);
         self
     }
+
+    /// Skip a certain user.
+    pub fn skip_user(&mut self, user_id: u64) -> &mut Self {
+        self.skip_users
+            .get_or_insert_with(BTreeSet::new)
+            .insert(user_id);
+        self
+    }
+
+    /// Skip a set of users.
+    pub fn skip_users<I>(&mut self, iter: I) -> &mut Self
+    where
+        I: Iterator<Item = u64>,
+    {
+        self.skip_users
+            .get_or_insert_with(BTreeSet::new)
+            .extend(iter);
+        self
+    }
 }
 
 impl<'a> Endpoint for GroupMembers<'a> {
@@ -67,7 +89,8 @@ impl<'a> Endpoint for GroupMembers<'a> {
 
         params
             .push_opt("query", self.query.as_ref())
-            .extend(self.user_ids.iter().map(|&value| ("user_ids[]", value)));
+            .extend(self.user_ids.iter().map(|&value| ("user_ids[]", value)))
+            .extend(self.skip_users.iter().map(|&value| ("skip_users[]", value)));
 
         params
     }
@@ -137,6 +160,24 @@ mod tests {
             .group("group/subgroup")
             .user_id(1)
             .user_ids([1, 2].iter().copied())
+            .build()
+            .unwrap();
+        api::ignore(endpoint).query(&client).unwrap();
+    }
+
+    #[test]
+    fn endpoint_skip_users() {
+        let endpoint = ExpectedUrl::builder()
+            .endpoint("groups/group%2Fsubgroup/members")
+            .add_query_params(&[("skip_users[]", "1"), ("skip_users[]", "2")])
+            .build()
+            .unwrap();
+        let client = SingleTestClient::new_raw(endpoint, "");
+
+        let endpoint = GroupMembers::builder()
+            .group("group/subgroup")
+            .skip_user(1)
+            .skip_users([1, 2].iter().copied())
             .build()
             .unwrap();
         api::ignore(endpoint).query(&client).unwrap();